fix: 그래픽 업로드 수정
